Setup main route
